Add build platform SO per platform
