+++ Runtime/BuildClass.cs
@@ -1,7 +1,4 @@
-﻿using GameWorkstore.Patterns;
-using System.IO;
-using UnityEditor;
-using UnityEditor.Build.Reporting;
+﻿using UnityEditor;
 using UnityEngine;
 
 namespace GameWorkstore.Automation
@@ -11,518 +8,61 @@ namespace GameWorkstore.Automation
         public static void BuildAndroid()
         {
             var buildScript = GetBuildScript();
-            BuildAndroid(buildScript);
-        }
-
-        public static void BuildAndroid(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-            
-            //PackageName
-            if (buildScript.BuildAndroid.UsePackageName && !string.IsNullOrEmpty(buildScript.BuildAndroid.PackageName))
-            {
-                PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.Android, buildScript.BuildAndroid.PackageName);
-            }
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var gameVersion, out var bundleVersion);
-                PlayerSettings.bundleVersion = gameVersion;
-                PlayerSettings.Android.bundleVersionCode = bundleVersion;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildAndroid.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Android,
-                    buildScript.BuildAndroid.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Android, buildScript.BuildAndroid.ScriptingBackend);
-
-            //Keystore Config
-            SetAndroidSignCredentials(buildScript);
-
-            if (buildScript.BuildAndroid.BuildAAB)
-            {
-                //Make
-                const string subPathAab = "Build/Android/AAB/";
-                MakeDirectory(subPathAab);
-
-                var buildOptions = new BuildPlayerOptions
-                {
-                    scenes = buildScript.BuildAndroid.GetScenes(),
-                    locationPathName = subPathAab + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".aab",
-                    target = BuildTarget.Android,
-                    options = GetOptions(buildScript.BuildAndroid)
-                };
-
-                EditorUserBuildSettings.buildAppBundle = true;
-                ProcessReport(BuildPipeline.BuildPlayer(buildOptions));
-            }
-
-            if (buildScript.BuildAndroid.BuildAPK)
-            {
-                //Make
-                const string subPathAPK = "Build/Android/APK/";
-                MakeDirectory(subPathAPK);
-
-                var buildOptions = new BuildPlayerOptions
-                {
-                    scenes = buildScript.BuildAndroid.GetScenes(),
-                    locationPathName = subPathAPK + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".apk",
-                    target = BuildTarget.Android,
-                    options = GetOptions(buildScript.BuildAndroid)
-                };
-
-                EditorUserBuildSettings.buildAppBundle = false;
-                ProcessReport(BuildPipeline.BuildPlayer(buildOptions));
-            }
+            buildScript.AndroidBuildPlatform.Build();
         }
 
         public static void BuildIOS()
         {
             var buildScript = GetBuildScript();
-            BuildIOS(buildScript);
-        }
-
-        public static void BuildIOS(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //PackageName
-            if (buildScript.BuildIOS.UsePackageName && !string.IsNullOrEmpty(buildScript.BuildIOS.PackageName))
-            {
-                PlayerSettings.SetApplicationIdentifier(BuildTargetGroup.iOS, buildScript.BuildIOS.PackageName);
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildIOS.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.iOS,
-                    buildScript.BuildIOS.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var gameVersion, out var bundleVersion);
-                PlayerSettings.bundleVersion = gameVersion;
-                PlayerSettings.iOS.buildNumber = bundleVersion.ToString();
-            }
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildIOS.GetScenes(),
-                locationPathName = "Build/iOS/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_"),
-                target = BuildTarget.iOS,
-                options = GetOptions(buildScript.BuildIOS)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildIOS.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.IOSBuildPlatform.Build();
         }
 
         public static void BuildWindows()
         {
             var buildScript = GetBuildScript();
-            BuildWindows(buildScript);
-        }
-
-        public static void BuildWindows(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildWindows.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Standalone,
-                    buildScript.BuildWindows.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildScript.BuildWindows.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildWindows.GetScenes(),
-                locationPathName = "Build/Windows/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".exe",
-                target = BuildTarget.StandaloneWindows64,
-                options = GetOptions(buildScript.BuildWindows)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildWindows.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.WindowsBuildPlatform.Build();
         }
 
         public static void BuildMacOS()
         {
             var buildScript = GetBuildScript();
-            BuildMacOS(buildScript);
-        }
-
-        public static void BuildMacOS(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildMacOS.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Standalone,
-                    buildScript.BuildMacOS.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildScript.BuildMacOS.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildMacOS.GetScenes(),
-                locationPathName = "Build/MacOS/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".app",
-                target = BuildTarget.StandaloneOSX,
-                options = GetOptions(buildScript.BuildMacOS)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildMacOS.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.MacOSBuildPlatform.Build();
         }
 
         public static void BuildLinux()
         {
             var buildScript = GetBuildScript();
-            BuildLinux(buildScript);
-        }
-        
-        public static void BuildLinux(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildLinux.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Standalone,
-                    buildScript.BuildLinux.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildScript.BuildLinux.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildLinux.GetScenes(),
-                locationPathName = "Build/Linux/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_"),
-                target = BuildTarget.StandaloneLinux64,
-                options = GetOptions(buildScript.BuildLinux)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildLinux.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.LinuxBuildPlatform.Build();
         }
 
         public static void BuildServerWindows()
         {
             var buildScript = GetBuildScript();
-            BuildServerWindows(buildScript);
+            buildScript.ServerWindowsBuildPlatform.Build();
         }
         
-        public static void BuildServerWindows(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildServerWindows.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Server,
-                    buildScript.BuildServerWindows.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildScript.BuildServerWindows.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildServerWindows.GetScenes(),
-                locationPathName = "Build/GameServerWindows/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".exe",
-                target = BuildTarget.StandaloneWindows64,
-                subtarget = (int)StandaloneBuildSubtarget.Server,
-                options = GetOptions(buildScript.BuildServerWindows)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildServerWindows.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
-        }
         public static void BuildServerLinux()
         {
             var buildScript = GetBuildScript();
-            BuildServerLinux(buildScript);
-        }
-        
-        public static void BuildServerLinux(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildServerLinux.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Server,
-                    buildScript.BuildServerLinux.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildScript.BuildServerLinux.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildServerLinux.GetScenes(),
-                locationPathName = "Build/GameServerLinux/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_"),
-                target = BuildTarget.StandaloneLinux64,
-                subtarget = (int)StandaloneBuildSubtarget.Server,
-                options = GetOptions(buildScript.BuildServerLinux)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildServerLinux.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.ServerLinuxBuildPlatform.Build();
         }
 
         public static void BuildServerMacOS()
         {
             var buildScript = GetBuildScript();
-            BuildServerMacOS(buildScript);
-        }
-        
-        public static void BuildServerMacOS(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildServerMacOS.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.Server,
-                    buildScript.BuildServerMacOS.ScriptDefinitions.Definitions
-                );
-            }
-
-            var buildConfig = buildScript.BuildServerMacOS;
-            //Backend
-            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, buildConfig.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildConfig.GetScenes(),
-                locationPathName = "Build/GameServerMacOS/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".app",
-                target = BuildTarget.StandaloneOSX,
-                subtarget = (int)StandaloneBuildSubtarget.Server,
-                options = GetOptions(buildConfig)
-            };
-
-            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
-            if (ProcessReportIsSuccess(buildReport))
-            {
-                CopyAdditionalFolders(buildOptions, buildScript.BuildServerMacOS.AdditionalFolders);
-            }
-            ProcessReport(buildReport);
+            buildScript.ServerMacOSBuildPlatform.Build();
         }
 
         public static void BuildUWP()
         {
             var buildScript = GetBuildScript();
-            BuildUWP(buildScript);
-        }
-        
-        public static void BuildUWP(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-                if (System.Version.TryParse(version, out var iVersion))
-                {
-                    PlayerSettings.WSA.packageVersion = iVersion;
-                }
-                else
-                {
-                    DebugMessege.Log("Failed to parse game version.", DebugLevel.ERROR);
-                }
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildUWP.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.WindowsStoreApps,
-                    buildScript.BuildUWP.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Backend
-            //Always IL2CPP
-            //PlayerSettings.SetScriptingBackend(BuildTargetGroup.WSA, buildScript.BuildUWP.ScriptingBackend);
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildUWP.GetScenes(),
-                locationPathName = "Build/UWP/",
-                target = BuildTarget.WSAPlayer,
-                options = GetOptions(buildScript.BuildUWP)
-            };
-
-            ProcessReport(BuildPipeline.BuildPlayer(buildOptions));
+            buildScript.UWPBuildPlatform.Build();
         }
 
         public static void BuildWebGL()
         {
             var buildScript = GetBuildScript();
-            BuildWebGL(buildScript);
-        }
-        
-        public static void BuildWebGL(BuildScript buildScript)
-        {
-            if (!Validate(buildScript)) return;
-
-            //Version
-            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
-            {
-                Version(out var version, out _);
-                PlayerSettings.bundleVersion = version;
-            }
-
-            //SetScriptDefinitions
-            if (buildScript.BuildWebGL.UseCustomScriptDefinitions)
-            {
-                PlayerSettings.SetScriptingDefineSymbols(
-                    UnityEditor.Build.NamedBuildTarget.WebGL,
-                    buildScript.BuildWebGL.ScriptDefinitions.Definitions
-                );
-            }
-
-            //Options
-            var buildOptions = new BuildPlayerOptions
-            {
-                scenes = buildScript.BuildWebGL.GetScenes(),
-                locationPathName = "Build/WebGL/",
-                target = BuildTarget.WebGL,
-                options = GetOptions(buildScript.BuildWebGL)
-            };
-
-            ProcessReport(BuildPipeline.BuildPlayer(buildOptions));
-        }
-
-        public static void Version(out string gameVersion, out int bundleVersion)
-        {
-            var g = Arg("-gameversion");
-            if (string.IsNullOrEmpty(g))
-			{
-                g = Arg("-buildVersion");
-			}
-            gameVersion = string.IsNullOrEmpty(g)? PlayerSettings.bundleVersion : g;
-
-            var bv = Arg("-gamebundleversion");
-			if (string.IsNullOrEmpty(bv))
-			{
-                bv = Arg("-androidVersionCode");
-                if (string.IsNullOrEmpty(bv))
-                {
-                    Debug.LogWarning("-bundleversion is legacy and conflicts with Unity iOS bundleversion arg. Use -gamebundleversion instead.");
-                    bv = Arg("-bundleversion");
-                }
-            }
-            bundleVersion = int.TryParse(bv, out bundleVersion)? bundleVersion : 1;
-
-            Debug.Log("GameVersion:[" + gameVersion + "][" + bundleVersion + "]");
+            buildScript.WebGLBuildPlatform.Build();
         }
 
         public static string Arg(string argument)
@@ -536,18 +76,6 @@ namespace GameWorkstore.Automation
 
             }
             return null;
-        }
-
-        public static void MakeDirectory(string relativePath)
-        {
-            var path = GetProjectPath() + relativePath;
-            if (Directory.Exists(path)) return;
-            Directory.CreateDirectory(path);
-        }
-
-        public static string GetProjectPath()
-        {
-            return Application.dataPath.Substring(0, Application.dataPath.Length - 6);
         }
 
         public static BuildScript GetBuildScript(bool logPath = false)
@@ -574,111 +102,6 @@ namespace GameWorkstore.Automation
                 }
             }
             return null;
-        }
-
-        private static BuildOptions GetOptions(BuildBase buildBase, BuildOptions baseOptions = BuildOptions.None)
-        {
-            return (buildBase.Development ? BuildOptions.Development : BuildOptions.None) | baseOptions;
-        }
-
-        public static bool ProcessReportIsSuccess(BuildReport buildReport)
-        {
-            return buildReport.summary.result == BuildResult.Succeeded;
-        }
-
-        public static void ProcessReport(BuildReport buildReport)
-        {
-            if (UnityEditorInternal.InternalEditorUtility.isHumanControllingUs) return;
-            if (ProcessReportIsSuccess(buildReport)) return;
-            EditorApplication.Exit(1);
-        }
-
-        public static bool Validate(BuildScript buildScript)
-        {
-            if (buildScript == null)
-            {
-                DebugMessege.Log("Cannot build without a BuildScript.", DebugLevel.ERROR);
-                return false;
-            }
-            if (string.IsNullOrEmpty(buildScript.GameName))
-            {
-                DebugMessege.Log("GameName cannot be null or empty.", DebugLevel.ERROR);
-                return false;
-            }
-            if(buildScript.GameVersionWriterConfig.Enabled && string.IsNullOrEmpty(buildScript.GameVersionWriterConfig.Namespace))
-            {
-                DebugMessege.Log("GameVersionWriter namespace cannot be null or empty.", DebugLevel.ERROR);
-                return false;
-            }
-            return true;
-        }
-
-        [MenuItem("Help/Automation/SetAndroidCredentials")]
-        public static void SetAndroidCredentials()
-        {
-            var buildScript = GetBuildScript();
-            SetAndroidSignCredentials(buildScript);
-        }
-
-        public static void SetAndroidSignCredentials(BuildScript buildScript)
-        {
-            if(buildScript == null) return;
-            if (!buildScript.BuildAndroid.UseKeystore) return;
-            PlayerSettings.Android.useCustomKeystore = true;
-            PlayerSettings.Android.keystoreName = GetProjectPath() + buildScript.BuildAndroid.KeyStoreSettings.KeystorePath;
-            PlayerSettings.Android.keystorePass = buildScript.BuildAndroid.KeyStoreSettings.KeystorePassword;
-            PlayerSettings.Android.keyaliasName = buildScript.BuildAndroid.KeyStoreSettings.AliasName;
-            PlayerSettings.Android.keyaliasPass = buildScript.BuildAndroid.KeyStoreSettings.AliasPassword;
-        }
-
-        private static void CopyAdditionalFolders(BuildPlayerOptions buildOptions, string[] additionalFolders)
-        {
-            string targetRoot = GetFolder(buildOptions);
-            if (string.IsNullOrEmpty(targetRoot)) return;
-
-            foreach(var additionalFolder in additionalFolders)
-            {
-                var source = Path.Combine(GetProjectPath(), additionalFolder);
-                var target = Path.Combine(targetRoot, additionalFolder);
-                if (Directory.Exists(target))
-                {
-                    Directory.Delete(target, true);
-                }
-                CopyFilesRecursively(source, target);
-            }
-        }
-
-        private static void CopyFilesRecursively(string sourcePath, string targetPath)
-        {
-            Directory.CreateDirectory(targetPath);
-            //Now Create all of the directories
-            foreach (string dirPath in Directory.GetDirectories(sourcePath, "*", SearchOption.AllDirectories))
-            {
-                Directory.CreateDirectory(dirPath.Replace(sourcePath, targetPath));
-            }
-
-            //Copy all the files & Replaces any files with the same name
-            foreach (string newPath in Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories))
-            {
-                File.Copy(newPath, newPath.Replace(sourcePath, targetPath), true);
-            }
-        }
-
-        public static string GetFolder(BuildPlayerOptions buildOptions)
-        {
-            var path = Path.Combine(GetProjectPath(), buildOptions.locationPathName);
-            switch (buildOptions.target)
-            {
-                case BuildTarget.StandaloneWindows:
-                case BuildTarget.StandaloneWindows64:
-                case BuildTarget.StandaloneOSX:
-                case BuildTarget.StandaloneLinux64:
-                    return Directory.GetParent(path).FullName;
-                case BuildTarget.iOS:
-                    return path;
-                default:
-                    return string.Empty;
-            }
         }
     }
 }
