+++ Runtime/Plataforms.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a48041de7776ff34f81767014654eae7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
