+++ Runtime/Plataforms/AndroidBuildPlatform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c3b121882abb1b45b54827a794e1850
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: efbe9a8fe7b3f6648956910062c56bd7, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
