+++ Runtime/Plataforms/LinuxBuildPlatform.cs
@@ -1,0 +1,54 @@
+using System;
+using System.Linq;
+using System.IO;
+using UnityEditor;
+using UnityEngine;
+using UnityEditor.Build.Reporting;
+using GameWorkstore.Patterns;
+
+namespace GameWorkstore.Automation
+{
+    [CreateAssetMenu(fileName = nameof(LinuxBuildPlatform), menuName = "Automation/" + nameof(LinuxBuildPlatform))]
+    public class LinuxBuildPlatform : StandaloneBuildPlatform
+    {
+        public override void Build()
+        {
+            if (!Validate(buildScript)) return;
+
+            //Version
+            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
+            {
+                Version(out var version, out _);
+                PlayerSettings.bundleVersion = version;
+            }
+
+            //SetScriptDefinitions
+            if (UseCustomScriptDefinitions)
+            {
+                PlayerSettings.SetScriptingDefineSymbols(
+                    UnityEditor.Build.NamedBuildTarget.Standalone,
+                    ScriptDefinitions.Definitions
+                );
+            }
+
+            //Backend
+            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, ScriptingBackend);
+
+            //Options
+            var buildOptions = new BuildPlayerOptions
+            {
+                scenes = GetScenes(),
+                locationPathName = "Build/Linux/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_"),
+                target = BuildTarget.StandaloneLinux64,
+                options = GetOptions()
+            };
+
+            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
+            if (ProcessReportIsSuccess(buildReport))
+            {
+                CopyAdditionalFolders(buildOptions, AdditionalFolders);
+            }
+            ProcessReport(buildReport);
+        }
+    }
+}
