+++ Runtime/Plataforms/ServerWindowsBuildPlatform.cs
@@ -1,0 +1,55 @@
+using System;
+using System.Linq;
+using System.IO;
+using UnityEditor;
+using UnityEngine;
+using UnityEditor.Build.Reporting;
+using GameWorkstore.Patterns;
+
+namespace GameWorkstore.Automation
+{
+    [CreateAssetMenu(fileName = nameof(ServerWindowsBuildPlatform), menuName = "Automation/" + nameof(ServerWindowsBuildPlatform))]
+    public class ServerWindowsBuildPlatform : StandaloneBuildPlatform
+    {
+        public override void Build()
+        {
+            if (!Validate(buildScript)) return;
+
+            //Version
+            if (!UnityEditorInternal.InternalEditorUtility.isHumanControllingUs)
+            {
+                Version(out var version, out _);
+                PlayerSettings.bundleVersion = version;
+            }
+
+            //SetScriptDefinitions
+            if (UseCustomScriptDefinitions)
+            {
+                PlayerSettings.SetScriptingDefineSymbols(
+                    UnityEditor.Build.NamedBuildTarget.Server,
+                    ScriptDefinitions.Definitions
+                );
+            }
+
+            //Backend
+            PlayerSettings.SetScriptingBackend(BuildTargetGroup.Standalone, ScriptingBackend);
+
+            //Options
+            var buildOptions = new BuildPlayerOptions
+            {
+                scenes = GetScenes(),
+                locationPathName = "Build/GameServerWindows/" + buildScript.GameName.ToLower().Replace(" ","_") + "_" + PlayerSettings.bundleVersion.Replace(".","_") + ".exe",
+                target = BuildTarget.StandaloneWindows64,
+                subtarget = (int)StandaloneBuildSubtarget.Server,
+                options = GetOptions()
+            };
+
+            var buildReport = BuildPipeline.BuildPlayer(buildOptions);
+            if (ProcessReportIsSuccess(buildReport))
+            {
+                CopyAdditionalFolders(buildOptions, AdditionalFolders);
+            }
+            ProcessReport(buildReport);
+        }
+    }
+}
