+++ Runtime/Plataforms/ServerWindowsBuildPlatform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 94735df9b77cfac49ab70890372be4d5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: efbe9a8fe7b3f6648956910062c56bd7, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
