+++ Runtime/Plataforms/UWPBuildPlatform.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d0fc8982c7e33b4091b2a391b91e921
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: efbe9a8fe7b3f6648956910062c56bd7, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
